refactor(checker): split module namespace exports
